requirements.txt
streamlit
python-dotenv
ibm-watson-machine-learning